+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - One unified configuration command
  - Editable original message from the withdrawal command
  - ID storage with addid command
+ - Auto-deletion of non-command messages in banking channel
+ - Banking channel to config
 
 ## Changed
  - Withdrawal command to have a specific on reaction loop instead of a general loop